🚚 appディレクトリ内のファイルを移動し構成を変更
