--- src/app/(pages)/profile/page.tsx
@@ -1,9 +1,0 @@
-import { ProfileCard } from "@/app/components/ProfileCard";
-
-export default function Page() {
-  return (
-    <div>
-      <ProfileCard />
-    </div>
-  );
-}
