+++ src/app/(pages)/records/add/loading.tsx
@@ -1,0 +1,6 @@
+import { BoxSpinner } from "@/app/common/components/BoxSpinner";
+import React from "react";
+
+export default function Loading() {
+  return <BoxSpinner />;
+}
