+++ src/app/(pages)/records/components/RecordList/Price.tsx
