+++ src/app/(pages)/records/components/RecordList/Record.tsx
