+++ src/app/(pages)/records/components/RecordList/RecordList.tsx
@@ -1,4 +1,4 @@
-import { Record } from "./elements/Record";
+import { Record } from "./Record";
 import { Card, Typography } from "@/app/common/lib/material-tailwind";
 import { fetchData } from "@/app/store/api/server/fetchData";
 import { RecordsType, zRecords } from "@/types";
