+++ src/app/(pages)/records/components/RecordList/index.ts
@@ -1,0 +1,1 @@
+export * from "./RecordList";
