+++ src/app/common/icons/CloseButton.tsx
