+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import "./globals.css";
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
-import { Header } from "@/app/components/Header";
+import { Header } from "@/app/common/components/Header";
 
 const inter = Inter({ subsets: ["latin"] });
